requirements.txt
pymorphy2
mastodon.py
